Make app responsive
