+++ src/components/app/index.ts
@@ -1,5 +1,7 @@
 export * from "./AppContent/AppContent";
+export * from "./AppMobileNav/AppMobileNav";
 export * from "./AppSidebar/AppSidebar";
+export * from "./AppSidebarHeader/AppSidebarHeader";
 export * from "./Logo/Logo";
 export * from "./PageHeader/PageHeader";
 export * from "./PasswordStrength/PasswordStrength";
