+++ src/components/settings/index.ts
@@ -9,3 +9,4 @@ export * from "./EditThemeSetting/EditThemeSetting";
 export * from "./FormResponsesList/FormResponsesList";
 export * from "./SettingsGroup/SettingsGroup";
 export * from "./SettingsItem/SettingsItem";
+export * from "./SettingsNav/SettingsNav";
