+++ src/routes/_authenticated.tsx
@@ -1,3 +1,5 @@
+import { AppMobileNav } from "@/components/app";
+import { useIsMobile } from "@/utils/useIsMobile";
 import { Outlet, createFileRoute, redirect } from "@tanstack/react-router";
 import { Authenticated } from "convex/react";
 
@@ -10,9 +12,14 @@ export const Route = createFileRoute("/_authenticated")({
 });
 
 function AuthenticatedRoute() {
+  const isMobile = useIsMobile();
+
   return (
     <Authenticated>
-      <Outlet />
+      <main className="grid grid-rows-[1fr_auto] h-screen min-h-0">
+        <Outlet />
+        {isMobile && <AppMobileNav />}
+      </main>
     </Authenticated>
   );
 }
