+++ src/routes/_authenticated/_home/birth-certificate.index.tsx
@@ -1,4 +1,3 @@
-import { AppContent } from "@/components/app";
 import { JurisdictionInterstitial } from "@/components/quests";
 import { createFileRoute, redirect } from "@tanstack/react-router";
 
@@ -19,9 +18,5 @@ export const Route = createFileRoute(
 });
 
 function RouteComponent() {
-  return (
-    <AppContent>
-      <JurisdictionInterstitial type="birthCertificate" />
-    </AppContent>
-  );
+  return <JurisdictionInterstitial type="birthCertificate" />;
 }
