+++ src/routes/_authenticated/settings/account.tsx
@@ -9,6 +9,7 @@ import {
   EditThemeSetting,
   SettingsGroup,
 } from "@/components/settings";
+import { useIsMobile } from "@/utils/useIsMobile";
 import { api } from "@convex/_generated/api";
 import { createFileRoute } from "@tanstack/react-router";
 import { useQuery } from "convex/react";
@@ -19,10 +20,14 @@ export const Route = createFileRoute("/_authenticated/settings/account")({
 
 function SettingsAccountRoute() {
   const user = useQuery(api.users.getCurrent);
+  const isMobile = useIsMobile();
 
   return (
     <>
-      <PageHeader title="Account" />
+      <PageHeader
+        title="Account"
+        backLink={isMobile ? { to: "/settings" } : undefined}
+      />
       {user === undefined ? (
         "Loading..."
       ) : user === null ? (
