+++ src/routes/_authenticated/settings/index.tsx
@@ -1,7 +1,25 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { PageHeader } from "@/components/app";
+import { SettingsNav } from "@/components/settings";
+import { useIsMobile } from "@/utils/useIsMobile";
+import { createFileRoute, useNavigate } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/_authenticated/settings/")({
-  beforeLoad: () => {
-    throw redirect({ to: "/settings/account" });
-  },
+  component: SettingsIndexRoute,
 });
+
+function SettingsIndexRoute() {
+  const navigate = useNavigate();
+  const isMobile = useIsMobile();
+
+  if (!isMobile) navigate({ to: "/settings/account" });
+
+  return (
+    <>
+      <PageHeader
+        title="Settings"
+        backLink={isMobile ? { to: "/" } : undefined}
+      />
+      <SettingsNav />
+    </>
+  );
+}
